add fungsi update

+++ app/Controllers/SuratMasuk.php
@@ -136,15 +136,15 @@ class SuratMasuk extends BaseController
 
     public function edit($id)
     {
-        $ids = session()->get('id_user');
+        $ids = session()->get('id');
         $data = array(
             'titlebar' => 'Surat Masuk',
             'title' => 'Edit Surat Masuk',
             'isi' => 'master/surat-masuk/edit',
             'validation' => \Config\Services::validation(),
-            'data' => $this->suratmasukModel->where('id', $id)->where('id_user', $ids)->first(),
+            'data' => $this->suratmasukModel->where('id =', $id)->where('id_user =', $ids)->first(),
         );
-        return view('admin/layout/wrapper', $data);
+        return view('layout/wrapper', $data);
     }
     public function update($id)
     {
@@ -183,9 +183,8 @@ class SuratMasuk extends BaseController
                 ]
             ],
             'file_surat' => [
-                'rules' => 'uploaded[file_surat]|mime_in[file_surat,application/pdf]|max_size[file_surat,2000]',
+                'rules' => 'mime_in[file_surat,application/pdf]|max_size[file_surat,2000]',
                 'errors' => [
-                    'uploaded' => 'FIle Surat harus diupload.',
                     'mime_in' => 'Extension file yang diperbolehkan .pdf',
                     'max_size' => 'Ukuran File maksimal 2MB.'
                 ]
@@ -205,25 +204,31 @@ class SuratMasuk extends BaseController
         if ($surat->getError() == 4) {
             $r = $this->suratmasukModel->find($id);
             $fileNamesurat = $r['file'];
+        } else {
+            $fileNamesurat = $surat->getRandomName();
+            //move file
+            $surat->move(ROOTPATH . 'public/media/surat-masuk/', $fileNamesurat);
+            //if file found then replace file
+            $f = $this->suratmasukModel->find($id);
+            $replacesurat = $f['file'];
+            if (file_exists(ROOTPATH . 'public/media/surat-masuk/' . $replacesurat)) {
+                unlink(ROOTPATH . 'public/media/surat-masuk/' . $replacesurat);
+            }
         }
         if ($lampiran->getError() == 4) {
             $r = $this->suratmasukModel->find($id);
             $fileNamelampiran = $r['lampiran'];
         } else {
-            $fileNamesurat = $surat->getRandomName();
             $fileNamelampiran = $lampiran->getRandomName();
             //move file
-            $surat->move(ROOTPATH . 'public/media/surat-masuk/', $fileNamesurat);
             $lampiran->move(ROOTPATH . 'public/media/lampiran/', $fileNamelampiran);
             //if file found then replace file
             $f = $this->suratmasukModel->find($id);
-            $replacesurat = $f['file'];
-            $replacelampiran = $f['surat'];
-            if (file_exists(ROOTPATH . 'public/media/surat-masuk/' . $replacesurat)) {
-                unlink(ROOTPATH . 'public/media/surat-masuk/' . $replacesurat);
-            }
+            $replacelampiran = $f['lampiran'];
             if (file_exists(ROOTPATH . 'public/media/lampiran/' . $replacelampiran)) {
-                unlink(ROOTPATH . 'public/media/lampiran/' . $replacelampiran);
+                if ($replacelampiran != null) {
+                    unlink(ROOTPATH . 'public/media/lampiran/' . $replacelampiran);
+                }
             }
         }
         $data = [
@@ -232,7 +237,7 @@ class SuratMasuk extends BaseController
             'sifat_surat'    => $this->request->getPost('sifat'),
             'kategori_surat' => $this->request->getPost('kategori'),
             'perihal'        => $this->request->getPost('perihal'),
-            'asal_surat'           => $this->request->getPost('asal'),
+            'asal_surat'     => $this->request->getPost('asal'),
             'file'           => $fileNamesurat,
             'lampiran'       => $fileNamelampiran,
         ];
